refactor(doctor): split check groups into focused files
Move check functions out of the monolithic doctor.go into three
dedicated files by concern:

- checks_build.go: checkStaleBuildArtifacts, checkBuildState,
  clientBinaryIssue, fleetStateIssue, checkCacheIntegrity
- checks_env.go: checkToolchainConsistency, checkDiskSpace,
  checkGitState, dockerNotInstalledDiagnostic
- checks_infra.go: checkAWSCredentialExpiry, checkDockerDaemon,
  checkDockerfileSecurity, lintResultToDiagnostic, downgradeRule,
  checkContainerImage

doctor.go retains only the Cmd, diagnostic type, and orchestration
helpers (runDoctor, printDiagnostics, countDiagnostics, etc.).
Reduces doctor.go complexity from 72 to ~15.

Tests split to match file boundaries; adds TestDowngradeRule (4 cases)
for previously untested function.
